migrations/models/3_20260412195104_add_image_urls_field.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `posts` ADD `image_urls` JSON NOT NULL COMMENT '图片URL列表';"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `posts` DROP COLUMN `image_urls`;"""
