requirements.txt
aerich==0.9.0
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.2
asyncclick==8.3.0.7
bcrypt==4.3.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
debugpy==1.8.20
dictdiffer==0.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Faker==40.1.2
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastapi-pagination==0.15.3
fastar==0.8.0
ffmpeg==1.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openpyxl==3.1.5
orjson==3.11.4
passlib==1.7.4
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMySQL==1.1.2
pypika-tortoise==0.6.3
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
starlette-sessions==0.3.0
tomlkit==0.13.2
tortoise-orm==0.25.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
